- fix IT tests - re-organized build.gradle a bit
